readFromFile flag in CP Solver, getting camel model from file, cdoservice, default values for current config arguments,

+++ utility-generator/src/main/java/eu/melodic/upperware/utilitygenerator/evaluator/EvaluatingUtils.java
@@ -95,7 +95,7 @@ public class EvaluatingUtils {
                 //int
                 case RAM:
                     log.debug("Creating getRamPredicate for value {}", var.getValue());
-                    predicates.add(getRamPredicate((long) var.getValue()));
+                    predicates.add(getRamPredicate((long) (int) var.getValue()));
                     break;
                 case CARDINALITY:
                     break;
